+++ src/router/index.ts
@@ -8,7 +8,10 @@ const routes: Array<RouteRecordRaw> = [
   {
     path: '/',
     name: 'Home',
-    component: Home
+    component: Home,
+    props: route => ({
+      modal: route.query.modal
+    })
   },
   {
     path: '/wallet',